Switch back to i32 for serving HTTP responses
ebird.ts:44 GET https://trogon.onrender.com/api/ebird-species?species_codes=cintin1,undtin1,horscr1,wfwduc1,bbwduc,musduc,specha3,pavpig2,plupig2,rudpig,blgdov1,greani1,smbani,litcuc2,squcuc1,compau,grepot1,rubher,whbher1,bltman1,blteme1,vereme1,glteme1,hoatzi1,soulap1,watjac1,yebter2,anhing,neocor,ruther1,cocher1,greegr,categr,strher,capher1,greibi1,bafibi1,bunibi1,gyhvul1,blchaw1,slchaw2,roahaw,fepowl,gnbtro1,rinkin1,amakin1,grnkin,chcpuf1,blfnun1,swwpuf1,whejac1,whcjac1,sccbar1,gilbar1,letara1,cheara1,scapic1,yetwoo2,linwoo1,rinwoo1,crcwoo2,spbwoo1,yehcar1,scspar1,cowpar1,blhpar1,blhpar4,brtpar1,scamac1,whepar2,greant1,blgant2,blcant3,silant1,plbwoo1,lobwoo1,strwoo2,butwoo1,stbwoo2,whbman1,witman2,putfru1,amaumb1,blttit1,yebfly1,sobtyr1,yectyr1,wilfly,cinatt1,brratt1,shcfly1,leskis1,grekis,bobfly1,socfly1,pirfly1,varfly,trokin,viojay1,bkcdon,whwswa1,houwre,bubwre1,pabthr1,blbthr1,ruboro1,sobcac1,yercac1,vefgra1,bkbwar,sibtan2,mactan1,bugtan,paltan1,turtan1,yebdac1,guitan1,chbsee1,grasal1 500 (Internal Server Error)

Feb 16 10:13:33 AM  thread '<unnamed>' panicked at 'error retrieving column taxonOrder: error deserializing column 4: cannot convert between the Rust type `f64` and the Postgres type `int4`', src/main.rs:1:1

This is a TODO: f64 is what works for --load-ebird-species

+++ src/models.rs
@@ -7,7 +7,7 @@ pub struct EbirdSpecies {
     pub comName: String,
     pub speciesCode: String,
     pub category: String,
-    pub taxonOrder: f64, // FIXME: i32/f64 confusion
+    pub taxonOrder: i32, // FIXME: i32/f64 confusion
     pub order: String,
     pub familyComName: Option<String>,
     pub familySciName: Option<String>,
